+++ jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctUndirectedGraph.java
@@ -62,6 +62,13 @@ import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
  * {@link Integer} as edge type. Please read the {@linkplain org.jgrapht.sux4j class documentation}
  * for more information.
  *
+ * <p>
+ * For convenience, and as a compromise with the approach of {@link SuccinctIntUndirectedGraph},
+ * this class provides methods {@link org.jgrapht.sux4j.SuccinctDirectedGraph#getEdgeFromIndex(int)
+ * getEdgeFromIndex()} and
+ * {@link org.jgrapht.sux4j.SuccinctDirectedGraph#getIndexFromEdge(it.unimi.dsi.fastutil.ints.IntIntPair)
+ * getIndexFromEdge()} that map bijectively the edge set into a contiguous set of longs.
+ *
  * @author Sebastiano Vigna
  * @see SuccinctIntUndirectedGraph
  */
@@ -211,13 +218,13 @@ public class SuccinctUndirectedGraph
      * @return the index associated with the edge, or &minus;1 if the edge is not part of the graph.
      * @see #getEdgeFromIndex(int)
      */
-    public int getIndexFromEdge(final IntIntSortedPair e)
+    public long getIndexFromEdge(final IntIntSortedPair e)
     {
         final int source = e.firstInt();
         final int target = e.secondInt();
         if (source < 0 || source >= n || target < 0 || target >= n)
             throw new IllegalArgumentException();
-        return (int) successors.indexOfUnsafe(((long) source << sourceShift) + target);
+        return successors.indexOfUnsafe(((long) source << sourceShift) + target);
     }
 
     /**
@@ -227,7 +234,7 @@ public class SuccinctUndirectedGraph
      * @return the pair with index {@code i}.
      * @see #getIndexFromEdge(IntIntSortedPair)
      */
-    public IntIntSortedPair getEdgeFromIndex(final int i)
+    public IntIntSortedPair getEdgeFromIndex(final long i)
     {
         if (i < 0 || i >= m)
             throw new IllegalArgumentException();
